refactored.  need to revert isCav changes in requests/ and components/
